Security added

+++ src/main/java/com/example/demo/repository/UserRepository.java
@@ -1,6 +1,7 @@
 package com.example.demo.repository;
 
 import com.example.demo.model.User;
+import com.example.demo.model.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,4 +9,4 @@ import org.springframework.stereotype.Repository;
  * Created by İbrahim Başar YARGICI at 28.09.2021
  */
 @Repository
-public interface UserRepository extends JpaRepository<User,Long> {}
+public interface UserRepository extends JpaRepository<User,Integer> {}
